+++ cornerstone-system/src/main/java/com/cornerstone/system/controller/SysDictController.java
@@ -36,8 +36,8 @@ public class SysDictController {
     @GetMapping("/type/page")
     @PreAuthorize("hasAuthority('system:dict:list')")
     public Result<Page<SysDictType>> pageType(
-            @RequestParam(name = "current", defaultValue = "1") long current,
-            @RequestParam(name = "size", defaultValue = "10") long size,
+            @RequestParam(name = "pageNum", defaultValue = "1") long current,
+            @RequestParam(name = "pageSize", defaultValue = "10") long size,
             @RequestParam(name = "dictName", required = false) String dictName,
             @RequestParam(name = "dictType", required = false) String dictType,
             @RequestParam(name = "status", required = false) String status) {
@@ -82,8 +82,8 @@ public class SysDictController {
     @GetMapping("/data/page")
     @PreAuthorize("hasAuthority('system:dict:list')")
     public Result<Page<SysDictData>> pageData(
-            @RequestParam(name = "current", defaultValue = "1") long current,
-            @RequestParam(name = "size", defaultValue = "10") long size,
+            @RequestParam(name = "pageNum", defaultValue = "1") long current,
+            @RequestParam(name = "pageSize", defaultValue = "10") long size,
             @RequestParam(name = "dictType", required = false) String dictType,
             @RequestParam(name = "dictLabel", required = false) String dictLabel,
             @RequestParam(name = "status", required = false) String status) {